Move `filterInfoCards`
